+++ couplet_composer/dependencies/sdl.py
@@ -466,87 +466,50 @@ def should_install(
             return False
 
 
-def install_dependency(
-    toolchain,
-    cmake_generator,
-    build_root,
-    dependencies_root,
-    version,
-    target,
-    host_system,
-    build_variant,
-    github_user_agent,
-    github_api_token,
-    opengl_version,
-    dry_run=None,
-    print_debug=None
-):
+def install_dependency(install_info, dry_run=None, print_debug=None):
     """
     Installs the dependency by downloading and possibly building
     it. Returns the path to the built dependency.
 
-    toolchain -- The toolchain object of the run.
-
-    cmake_generator -- The name of the generator that CMake
-    should use as the build system for which the build scripts
-    are generated.
-
-    build_root -- The path to the root directory that is used for
-    all created files and directories.
-
-    dependencies_root -- The root directory of the dependencies
-    for the current build target.
-
-    version -- The full version number of the dependency.
-
-    target -- The target system of the build represented by a
-    Target.
-
-    host_system -- The system this script is run on.
-
-    build_variant -- The build variant used to build the project.
-
-    github_user_agent -- The user agent used when accessing the
-    GitHub API.
-
-    github_api_token -- The GitHub API token that is used to
-    access the API.
-
-    opengl_version -- The version of OpenGL that is used.
+    install_info -- The object containing the install information
+    for this tool.
 
     dry_run -- Whether the commands are only printed instead of
     running them.
 
     print_debug -- Whether debug output should be printed.
     """
-    temp_dir = get_temporary_directory(build_root=build_root)
+    temp_dir = get_temporary_directory(build_root=install_info.build_root)
     dependency_temp_dir = os.path.join(temp_dir, "sdl")
 
     shell.makedirs(temp_dir, dry_run=dry_run, echo=print_debug)
     shell.makedirs(dependency_temp_dir, dry_run=dry_run, echo=print_debug)
 
     url = ("https://www.libsdl.org/release/SDL2-{version}.zip"
-           if host_system == get_windows_system_name()
+           if install_info.host_system == get_windows_system_name()
            else "https://www.libsdl.org/release/SDL2-{version}.tar.gz").format(
-        version=version
+        version=install_info.version
     )
     dest = os.path.join(
         dependency_temp_dir,
-        "sdl.zip" if host_system == get_windows_system_name() else "sdl.tar.gz"
+        "sdl.zip" if install_info.host_system == get_windows_system_name()
+        else "sdl.tar.gz"
     )
 
     http.stream(
         url=url,
         destination=dest,
-        host_system=host_system,
+        host_system=install_info.host_system,
         dry_run=dry_run,
         print_debug=print_debug
     )
     shell.tar(dest, dependency_temp_dir, dry_run=dry_run, echo=print_debug)
 
-    subdir = os.path.join(dependency_temp_dir, "SDL2-{}".format(version))
+    subdir = os.path.join(dependency_temp_dir, "SDL2-{}".format(
+        install_info.version
+    ))
 
-    if host_system == get_windows_system_name():
+    if install_info.host_system == get_windows_system_name():
         # _copy_visual_c_binaries(
         #     dependencies_root=dependencies_root,
         #     subdirectory=subdir,
@@ -554,21 +517,21 @@ def install_dependency(
         #     print_debug=print_debug
         # )
         _build_using_cmake(
-            toolchain=toolchain,
-            cmake_generator=cmake_generator,
-            dependencies_root=dependencies_root,
+            toolchain=install_info.toolchain,
+            cmake_generator=install_info.cmake_generator,
+            dependencies_root=install_info.dependencies_root,
             temporary_directory=temp_dir,
             subdirectory=subdir,
-            target=target,
-            host_system=host_system,
-            build_variant=build_variant,
+            target=install_info.target,
+            host_system=install_info.host_system,
+            build_variant=install_info.build_variant,
             dry_run=dry_run,
             print_debug=print_debug
         )
     else:
         _build(
-            toolchain=toolchain,
-            dependencies_root=dependencies_root,
+            toolchain=install_info.toolchain,
+            dependencies_root=install_info.dependencies_root,
             temporary_directory=temp_dir,
             subdirectory=subdir,
             dry_run=dry_run,