+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 ### Changed
 
 - The required parameters of the tool installation functions to one object instead of passing many different parameters.
+- The required parameters of the dependency installation functions to one object instead of passing many different parameters.
 - The version of Ninja to 1.10.0.
 
 ## [0.12.0] - 2020-04-25